requirements-tests.txt
-r requirements.txt
djangocms-helper
django-ckeditor
